backend/requirements-railway.txt
gunicorn==21.2.0
flask==2.3.3
flask-cors==4.0.0
pymongo==4.3.3
python-dotenv==1.0.0
PyJWT==2.8.0
Pillow==10.0.1
